Game Dev/The Coroner/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
alembic==1.15.2
pydantic==2.12.5
httpx==0.28.1
python-dotenv==1.2.2
jinja2==3.1.6
python-multipart==0.0.20
